requirements.txt
cv_bridge==1.16.2
message_filters==1.16.0
ml_collections==0.1.1
numpy==1.23.5
open3d==0.13.0
opencv_python==4.8.0.76
py3rosmsgs==1.18.2
PyYAML==6.0.1
PyYAML==6.0.1
rosnumpy==0.0.5.2
rospy==1.16.0
scipy==1.10.1
sensor_msgs==1.13.1
setuptools==68.0.0
torch==2.0.1+cu118
torchvision==0.15.2+cu118
